chore(web): remove manual type casts, lean on wagmi ABI inference
The ABI is exported from @ipe-gov/sdk as a const assertion, so wagmi
infers the return shape of getProposal / proposalCount correctly. Drop
the hand-written ProposalTuple and the 'as bigint' casts.

Also tighten the FHE SDK config: network is required by the SDK type,
and SepoliaConfig omits it, so always provide either the injected EIP-1193
provider (preferred) or a public Sepolia RPC URL as a fallback.

+++ apps/web/src/routes/proposals.index.tsx
@@ -26,7 +26,7 @@ function Proposals() {
     abi: UnlockConfidentialGovernorABI,
     functionName: 'proposalCount',
   })
-  const total = count ? Number(count as bigint) : 0
+  const total = count ? Number(count) : 0
   const ids = Array.from({ length: total }, (_, i) => total - i)
 
   return (
